define function so points can connect
the pointBounds function ensures that each specific point is held relatively close to the other points (simulating the inspiration that I got from FTL, where the ship can only travel a certain distance from each dot). it also provides a method for the function to regulate outliers in case a dot is far away from the others, which is important so that I simulate game-like conditions. the other function finds two ordered pairs from points (x1, y1) and (x2, y2) and iterates through the entire array so that each point but the first is connected to the point before it.